Use subprocesses instead of threads.

+++ collector.py
@@ -2,6 +2,20 @@ import os
 import sys
 import socket
 import json
+import subprocess
+import signal
+
+def signal_main(signum, frame):
+    print "Shutting down"
+    for name, process in processes.iteritems():
+        print "Stopping {0}".format(name)
+        process.terminate()
+    sink.close()
+    exit(0)
+
+def signal_plugin(signum, frame):
+    print "Exiting plugin {0}".format(instance)
+    m.shutdown = True
 
 if not __name__ == "__main__":
     print "The collector should not be imported as a module"
@@ -13,40 +27,50 @@ f = open("config", "r")
 conf = json.loads(f.read())
 socketpath = conf["socketpath"]
 
-# create a unix domain socket to be used as sink
-print "Preparing domain socket"
-sink = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM)
-try:
-    os.remove(socketpath)
-except:
-    pass
+# if passed an argument, we run that plugin instance
+if len(sys.argv) > 1:
+    instance = sys.argv[1]
+    if not instance in conf["plugins"]:
+        print "No configuration for {0}, exiting.".format(name)
+        os.exit(1)
 
-sink.bind(socketpath)
+    values = conf["plugins"][instance]
 
-# Prepare modules and plugins
-modules = {}
-plugins = {}
-for name, values in conf["plugins"].iteritems():
-    k = values["plugin"]
-    if not k in modules:
-        print "Importing {0}".format(k)
-        modules[k] = getattr(__import__('plugins.{0}'.format(k), globals(), locals(), [k], -1), k)
+    print "Importing {0}".format(values["plugin"])
+    modulename = values["plugin"]
+    module = getattr(__import__('plugins.{0}'.format(modulename), globals(), locals(), [modulename], -1), modulename)
+    m = module(socketpath, values)
 
-    print "Initializing {0}".format(name)
-    plugins[name] = modules[k](socketpath, values)
+    signal.signal(signal.SIGTERM, signal_plugin)
 
-#TODO see if stuff gets lost when starting takes a while
-for name, instance in plugins.iteritems():
-    print "Starting {0}".format(name)
-    instance.start()
+    print "Starting plugin {0}".format(instance)
+    while not m.shutdown:
+       m.run()
+    m.shutdown()
 
-try:
+
+else:
+    # create a unix domain socket to be used as sink
+    print "Preparing domain socket"
+    sink = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM)
+    try:
+        os.remove(socketpath)
+    except:
+        pass
+
+    sink.bind(socketpath)
+
+    # Prepare modules and plugins
+    processes = {}
+    #TODO see if stuff gets lost when starting takes a while
+    for name in conf["plugins"].keys():
+        print "Starting collector for plugin {0}".format(name)
+        processes[name] = subprocess.Popen([sys.executable, sys.argv[0], name])
+
+    #TODO regularly poll() the subprocesses
+
+    signal.signal(signal.SIGTERM, signal_main)
+
     while True:
         data, address = sink.recvfrom(4096) # hmm
         print >>sys.stderr, data
-except:
-    print "Shutting down"
-    for name, instance in plugins.iteritems():
-        print "Stopping {0}".format(name)
-        instance.shutdown = True
-    sink.close()
